% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_inpredictable.R
\name{get_current_inpredictable}
\alias{get_current_inpredictable}
\title{Returns current inpredictable nfl rankings (scraped)}
\usage{
get_current_inpredictable()
}
\value{
a dataframe containing current inpredictable nfl vegas power rankings
}
\description{
Returns current inpredictable nfl rankings (scraped)
}
