% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvt_game.R
\name{mvt_game}
\alias{mvt_game}
\title{Given game play-by-play data, returns finite time average point differential (i.e., time_avg_lead) for regulation time}
\usage{
mvt_game(game_pbp, team)
}
\arguments{
\item{game_pbp}{a play-by-play data file via nflfastR}

\item{team}{a team of interest}
}
\value{
a data frame containing game info and time average lead for regulation
}
\description{
Given game play-by-play data, returns finite time average point differential (i.e., time_avg_lead) for regulation time
}
