% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvt_season.R
\name{mvt_season}
\alias{mvt_season}
\title{For a given team and season, returns regulation time average lead for each game played in the season.}
\usage{
mvt_season(team_pbp_df, team)
}
\arguments{
\item{team_pbp_df}{a data frame with play-by-play data for season of interest}

\item{team}{a team of interest}
}
\value{
a data frame with time average lead results for each game played by team in a season
}
\description{
For a given team and season, returns regulation time average lead for each game played in the season.
}
