% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvt_sos.R
\name{mvt_sos}
\alias{mvt_sos}
\title{For a given team, returns strength of schedule measure based on opponent performance against other teams in regular season}
\usage{
mvt_sos(df, team_interest)
}
\arguments{
\item{df}{a long data frame with}

\item{team_interest}{a team of interest}
}
\value{
a data frame with a average strength of schedule measure for each opponent
}
\description{
For a given team, returns strength of schedule measure based on opponent performance against other teams in regular season
}
