% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_mvt_season.R
\name{nfl_mvt_season}
\alias{nfl_mvt_season}
\title{Returns long data frame with time average results for each team}
\usage{
nfl_mvt_season(year = 2021)
}
\arguments{
\item{year}{a year of interest}
}
\value{
a long data frame with time average results for each game for every team
}
\description{
Returns long data frame with time average results for each team
}
