% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_mvt_season_summary.R
\name{nfl_mvt_season_summary}
\alias{nfl_mvt_season_summary}
\title{Returns an aggregate summary of time average results with adjustments based on time average opponent quality metrics}
\usage{
nfl_mvt_season_summary(df)
}
\arguments{
\item{df}{a long data frame with time average results for each game for every team}
}
\value{
a data frame with aggregate time average results with adjustments based on time average opponent quality metrics
}
\description{
Returns an aggregate summary of time average results with adjustments based on time average opponent quality metrics
}
