% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_mvt_sos.R
\name{nfl_mvt_sos}
\alias{nfl_mvt_sos}
\title{Returns strength of schedule measure for all teams based on opponent performance against other teams in regular season}
\usage{
nfl_mvt_sos(df)
}
\arguments{
\item{df}{a long data frame with time average point and possession differential for each game in a season}
}
\value{
a data frame with a average strength of schedule measures for all teams
}
\description{
Returns strength of schedule measure for all teams based on opponent performance against other teams in regular season
}
