% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_reg_season_results.R
\name{nfl_reg_season_results}
\alias{nfl_reg_season_results}
\title{Returns long data frame documenting final result of each regular season game}
\usage{
nfl_reg_season_results(year = 2020)
}
\arguments{
\item{year}{a season of interest}
}
\value{
a long data frame with results of each game in a season
}
\description{
Returns long data frame documenting final result of each regular season game
}
