% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_reg_season_summary.R
\name{nfl_reg_season_summary}
\alias{nfl_reg_season_summary}
\title{Returns aggregate results for regular season (e.g., wins, losses, total points, pythagorean win expectation)}
\usage{
nfl_reg_season_summary(year = 2020)
}
\arguments{
\item{year}{a year of interest}
}
\value{
a data frame with aggregate regular season results
}
\description{
Returns aggregate results for regular season (e.g., wins, losses, total points, pythagorean win expectation)
}
