% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_reg_season_summary.R
\name{team_reg_season_summary}
\alias{team_reg_season_summary}
\title{For a given team and data frame of regular season results data, returns data frame with season summary results and pythagorean win expectation}
\usage{
team_reg_season_summary(df, team)
}
\arguments{
\item{df}{a data frame with regular season game results}

\item{team}{a team of interest}
}
\value{
a data frame with season summary results and pythagorean win expectation
}
\description{
For a given team and data frame of regular season results data, returns data frame with season summary results and pythagorean win expectation
}
